moving tests for hdf_io into appropriate location
