+++ tests/io/__init__.py
@@ -1,0 +1,2 @@
+from . import hdf_io
+__all__ = ['hdf_io']
